+++ src/GameDesignWizard.App/MainWindow.xaml.cs
@@ -1,6 +1,4 @@
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Input;
 using GameDesignWizard.App.ViewModels;
 using Microsoft.Win32;
 
@@ -17,51 +15,6 @@ public partial class MainWindow : Window
     {
         InitializeComponent();
         DataContext = viewModel;
-    }
-
-    private void AvailableTopic_MouseDoubleClick(object sender, MouseButtonEventArgs e)
-    {
-        TransferTopic(sender, add: true);
-        e.Handled = true;
-    }
-
-    private void SelectedTopic_MouseDoubleClick(object sender, MouseButtonEventArgs e)
-    {
-        TransferTopic(sender, add: false);
-        e.Handled = true;
-    }
-
-    private void AvailableTopic_KeyDown(object sender, KeyEventArgs e)
-    {
-        if (e.Key == Key.Enter)
-        {
-            TransferTopic(sender, add: true);
-            e.Handled = true;
-        }
-    }
-
-    private void SelectedTopic_KeyDown(object sender, KeyEventArgs e)
-    {
-        if (e.Key == Key.Enter)
-        {
-            TransferTopic(sender, add: false);
-            e.Handled = true;
-        }
-    }
-
-    private void TransferTopic(object sender, bool add)
-    {
-        if (sender is not ListBoxItem { DataContext: CatalogOptionViewModel topic }
-            || DataContext is not MainWindowViewModel viewModel)
-        {
-            return;
-        }
-
-        var command = add ? viewModel.AddTopicCommand : viewModel.RemoveTopicCommand;
-        if (command.CanExecute(topic))
-        {
-            command.Execute(topic);
-        }
     }
 
     private async void ImportCatalog_Click(object sender, RoutedEventArgs e)